bb-iced-widgets: Add circle_bar
- Just a circle with label in the middle.
- Used in flash finish pages.
