+++ bb-iced-widgets/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod circle_bar;
 pub mod progress_circle;
 
 pub fn progress_circle<T>(
@@ -7,4 +8,13 @@ pub fn progress_circle<T>(
     font: iced::Font,
 ) -> iced::widget::Canvas<progress_circle::ProgressCircle, T> {
     progress_circle::ProgressCircle::new(progress, thickness, color, font)
+}
+
+pub fn circle_bar<T>(
+    label: &'static str,
+    thickness: impl Into<f32>,
+    color: iced::Color,
+    font: iced::Font,
+) -> iced::widget::Canvas<circle_bar::CircleBar, T> {
+    circle_bar::CircleBar::new(label, thickness, color, font)
 }
